Update
Fixed scaling issue.
Changed the character limit before newline is added.
Changed so  drawn string is all uppercase, and removed string copy, as it made it redundant.
Removed  string_delete as it made wrapping behave not in the intended way.
Added text alignment fixing as it jumped around when the first broadcast popped.